refactor: inline staff dashboard stat card
