+++ app/routes/staff/index.tsx
@@ -9,7 +9,6 @@ import ShoppingCartIcon from "../../components/icons/lucide/shoppingCartIcon"
 import { getStaffDashboardData } from "../../usecases/getStaffDashboardData"
 import DailyOrdersTrendCard from "./-components/$dailyOrdersTrendCard"
 import StatusDistributionCard from "./-components/$statusDistributionCard"
-import DashboardStatCard from "./-components/dashboardStatCard"
 import Layout from "./-components/layout"
 import { buildDashboardStatCards } from "./-helpers/dashboardStatCards"
 
@@ -82,6 +81,28 @@ const QuickAccessCard: FC<QuickAccessCard> = ({
         </div>
       </div>
     </a>
+  )
+}
+
+type DashboardStatCardProps = {
+  title: string
+  value: string
+  description?: string
+}
+
+const DashboardStatCard = ({
+  title,
+  value,
+  description,
+}: DashboardStatCardProps) => {
+  return (
+    <div className="rounded-lg border bg-muted/30 p-4">
+      <p className="text-muted-fg text-sm">{title}</p>
+      <p className="mt-2 font-semibold text-2xl">{value}</p>
+      {description && (
+        <p className="mt-1 text-muted-fg text-xs">{description}</p>
+      )}
+    </div>
   )
 }
